+++ src/timeoff/prompts.py
@@ -1,64 +1,37 @@
 from datetime import datetime
 
 
-def date_prompt(prompt, default=None):
-    """Prompt to set a date."""
-    while True:
-        if default is None:
-            date = input(f"{prompt}: ")
-        else:
-            date = input(f"{prompt} [{default}]: ") or default
+def date_validator(*func_args):
+    func = func_args[0]
+    def wrapper(*args):
         try:
-            date = datetime.strptime(date, "%Y-%m-%d").date()
+            value = (datetime.strptime(args[0], "%Y-%m-%d").date(), )
         except ValueError:
-            print("Please enter a valid date")
-            continue
+            return "Please enter a valid date"
         else:
-            return date
+            return func(*value)
+    return wrapper
 
 
-def float_prompt(prompt, default=None):
-    """Prompt to set a float."""
-    while True:
-        if default is None:
-            value = input(f"{prompt}: ")
-        else:
-            value = input(f"{prompt} [{default}]: ") or default
+def float_validator(*func_args):
+    func = func_args[0]
+    def wrapper(*args):
         try:
-            value = float(value)
+            value = (float(args[0]), )
         except ValueError:
-            print("Please enter a valid number")
-            continue
+            return "Please enter a valid number"
         else:
-            return value
+            return func(*value)
+    return wrapper
 
 
-def int_prompt(prompt, default=None):
-    """Prompt to set an int."""
-    while True:
-        if default is None:
-            value = input(f"{prompt}: ")
-        else:
-            value = input(f"{prompt} [{default}]: ") or default
+def int_validator(*func_args):
+    func = func_args[0]
+    def wrapper(*args):
         try:
-            value = int(value)
+            value = (int(args[0]), )
         except ValueError:
-            print("Please enter a valid number")
-            continue
+            return "Please enter a valid number"
         else:
-            return value
-
-
-def str_prompt(prompt, default=None):
-    """Prompt to set a string."""
-    while True:
-        if default is None:
-            value = input(f"{prompt}: ")
-        else:
-            value = input(f"{prompt} [{default}]: ") or default
-
-        if not value:
-            print("Please enter a value")
-            continue
-
-        return value
+            return func(*value)
+    return wrapper