+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Use PyInquirer for prompts
+
 ## [0.0.2] - 2023-05-22
 
 ### Added